[DEV] update to last dev revision

+++ hardware/1Wire/1WireByOWFS.cpp
@@ -1,7 +1,7 @@
 #include "stdafx.h"
 #include "1WireByOWFS.h"
 #include "../../main/mainworker.h"
-#include "../main/Logger.h"
+#include "../../main/Logger.h"
 
 #include <fstream>
 #include <algorithm>
@@ -320,7 +320,7 @@ int C1WireByOWFS::GetVoltage(const _t1WireDevice& device,int unit) const
    case smart_battery_monitor:
       {
          static const std::string unitNames[] = { "VAD", "VDD", "vis" };
-         if (unit >= (sizeof(unitNames)/sizeof(unitNames[0])))
+         if (unit >= int((sizeof(unitNames)/sizeof(unitNames[0]))))
             return 0;
          fileName.append("/").append(unitNames[unit]);
          break;
